move check fw version test to api
